chore: remove MainStateStore & modify service file path
