+++ src/components/main/posts/Search.tsx
@@ -2,7 +2,7 @@
 import React, { useState } from "react";
 import Image from "next/image";
 import { useRecoilState } from "recoil";
-import { postSearchValue } from "@/store/MainStateStore";
+import { postSearchValue } from "@/store/post/PostStateStore";
 
 const Search = () => {
   const [searchValue, setSearchValue] = useRecoilState(postSearchValue);
