+++ src/hooks/useProfileInfo.tsx
@@ -1,12 +1,12 @@
 'use client';
 import { useSuspenseQuery } from "@tanstack/react-query";
-import { getUserIfo as getProfileInfoAPI } from "@/service/user";
+import { getUserIfo } from "@/service/user/user";
 import { ProfileInfo, ResponseBody } from "@/utils/type";
 
 export function useProfileInfo() {
   const { data } = useSuspenseQuery<ResponseBody<ProfileInfo>, Error>({
     queryKey: ['profileInfo'],
-    queryFn: getProfileInfoAPI
+    queryFn: getUserIfo
   });
 
   return data.data;
